integration/containers: Move options into a separated file
